feat[service]: rename of the folder from component to service
